Add stdio transport entry point alongside HTTP
Refactor src/index.ts to split MCP server creation and tool registration
(now in src/server.ts) from transport bootstrapping. With --stdio, the
server attaches StdioServerTransport instead of binding the Express HTTP
listener, and console.log is redirected to stderr so device-manager init
messages don't poison the JSON-RPC channel on stdout.

Default behavior (no flag) is unchanged — HTTP on port 3000.

Fixes #5
