+++ package.json
@@ -7,6 +7,7 @@
   "scripts": {
     "build": "tsc",
     "start": "node dist/index.js",
+    "start:stdio": "node dist/index.js --stdio",
     "dev": "tsc --watch"
   },
   "keywords": [
